Исправление ошибки. Вынесены функции для работы со стримами в отдельный файл.
